initial

+++ routes/notes.js
@@ -35,7 +35,7 @@ function formatTime(){
 	year = ''+year;
 
 	var date2 = new Date(parseInt(year), parseInt(month)-1, parseInt(date)); // Month value is zero-based
-	var options = { weekday: 'long' };
+	var options = { weekday: 'numeric' };
 	var weekday = date2.toLocaleDateString('en-US', options);
 	weekday = ''+weekday;
